Revert "Update field `nextTime` of struct `KlineResult` from unit second to millisecond."
This reverts commit 160f1e011e4c74e5e4ca95702974d1e58bde1d20.

+++ src/restapi/types.rs
@@ -26,7 +26,7 @@ pub struct KlineResult {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub v: Vec<f32>, // trading volume
     #[serde(rename = "nextTime", skip_serializing_if = "Option::is_none")]
-    pub nxt: Option<i64>,
+    pub nxt: Option<i32>,
 }
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, Apiv2Schema)]
